+++ package.json
@@ -20,7 +20,8 @@
     "react-dom": "^18.2.0",
     "react-instantsearch-hooks-web": "^6.40.1",
     "react-router-dom": "^6.8.2",
-    "react-scripts": "^5.0.1"
+    "react-scripts": "^5.0.1",
+    "typescript": "^5.0.2"
   },
   "browserslist": {
     "production": [
