+++ src/components/Panel.jsx
